UPDATE: refresh trusted device headers and add transaction parameters (TAP-2430) (#352)
* UPDATE: refresh trusted device headers and add transaction parameters (TAP-2430)

* UPDATE: refresh trusted device headers and add transaction parameters (TAP-2430)

* UPDATE: refresh trusted device headers and add transaction parameters (TAP-2430

* UPDATE: refresh trusted device headers and add transaction parameters (TAP-2430) CR and fix amount

* UPDATE: refresh trusted device headers and add transaction parameters (TAP-2430) CR fix named

* UPDATE: refresh trusted device headers and add transaction parameters (TAP-2430) CR fix summary screen for transfer to phone

* UPDATE: refresh trusted device headers and add transaction parameters (TAP-2430) CR

* Update Modules/Libraries/SANPLLibrary/Example/Tests/Mocks/PLHostProviderMock.swift cr


* UPDATE: refresh trusted device headers and add transaction parameters (TAP-2430) CR

* UPDATE: refresh trusted device headers and add transaction parameters (TAP-2430) CR

* UPDATE: refresh trusted device headers and add transaction parameters (TAP-2430) CR

* UPDATE: refresh trusted device headers and add transaction parameters (TAP-2430) CR

* UPDATE: refresh trusted device headers and add transaction parameters (TAP-2430) -create PLCryptography module CR

* UPDATE: refresh trusted device headers and add transaction parameters (TAP-2430) fix MobileTransferSummary  CR

* UPDATE: refresh trusted device headers and add transaction parameters (TAP-2430) fix MobileTransferSummary  CR

* UPDATE: refresh trusted device headers and add transaction parameter CR

* UPDATE: refresh trusted device headers and add transaction parameter CR

* UPDATE: refresh trusted device headers and add transaction parameter CR

* UPDATE: refresh trusted device headers and add transaction parameter CR

* UPDATE: refresh trusted device headers and add transaction parameter CR

* UPDATE: refresh trusted device headers and add transaction parameter CR

* UPDATE: refresh trusted device headers and add transaction parameter CR

* UPDATE: refresh trusted device headers and add transaction parameter CR

* UPDATE: refresh trusted device headers and add transaction parameter CR

* UPDATE: refresh trusted device headers and add transaction parameter CR

* UPDATE: refresh trusted device headers and add transaction parameter CR

* UPDATE: refresh trusted device headers and add transaction parameter CR

* UPDATE: refresh trusted device headers and add transaction parameter CR

+++ Modules/Features/BLIK/BLIK/Features/MobileTransfer/Confirmation/Models/MobileTransferSummary.swift
@@ -3,7 +3,7 @@ import SANLegacyLibrary
 import SANPLLibrary
 
 public struct MobileTransferSummary {
-    public let amount: Double
+    public let amount: Decimal
     public let currency: CurrencyType
     public let title: String
     public let accountName: String
